test(server): move domain lookup network cases to integration
